ACTIN-1605 Rename the parent interface LabEvaluationFunction and update all references.
