+++ algo/src/main/kotlin/com/hartwig/actin/algo/evaluation/laboratory/HasLimitedBilirubinPercentageOfTotal.kt
@@ -8,7 +8,7 @@ import com.hartwig.actin.datamodel.clinical.LabValue
 import java.time.LocalDate
 
 class HasLimitedBilirubinPercentageOfTotal(private val maxPercentage: Double, private val minValidDate: LocalDate) :
-    LabEvaluationFunction {
+    SingleLabValueEvaluationFunction {
 
     override fun evaluate(record: PatientRecord, labMeasurement: LabMeasurement, labValue: LabValue): Evaluation {
         val interpretation = LabInterpretation.interpret(record.labValues)
