+++ algo/src/main/kotlin/com/hartwig/actin/algo/evaluation/laboratory/HasSufficientLabValue.kt
@@ -11,7 +11,7 @@ import com.hartwig.actin.datamodel.clinical.LabValue
 
 class HasSufficientLabValue(
     private val minValue: Double, private val measurement: LabMeasurement, private val targetUnit: LabUnit
-) : LabEvaluationFunction {
+) : SingleLabValueEvaluationFunction {
 
     override fun evaluate(record: PatientRecord, labMeasurement: LabMeasurement, labValue: LabValue): Evaluation {
         val convertedValue = LabUnitConverter.convert(measurement, labValue, targetUnit)
