+++ algo/src/main/kotlin/com/hartwig/actin/algo/evaluation/laboratory/HasSufficientLabValueLLN.kt
@@ -8,7 +8,7 @@ import com.hartwig.actin.datamodel.PatientRecord
 import com.hartwig.actin.datamodel.algo.Evaluation
 import com.hartwig.actin.datamodel.clinical.LabValue
 
-class HasSufficientLabValueLLN(private val minLLNFactor: Double) : LabEvaluationFunction {
+class HasSufficientLabValueLLN(private val minLLNFactor: Double) : SingleLabValueEvaluationFunction {
 
     override fun evaluate(record: PatientRecord, labMeasurement: LabMeasurement, labValue: LabValue): Evaluation {
         val labValueString = labValue(labMeasurement, labValue.value, labValue.unit)
